backend/requirements.txt
flask
flask-cors
firebase-admin
requests
google-cloud-bigquery
google-cloud-aiplatform
google-genai
gunicorn
